% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-alternatives.R
\name{alternative_products}
\alias{alternative_products}
\title{Find alternative products for all products containing a certain active substance in PSM-V}
\usage{
alternative_products(name_active_ingredient)
}
\description{
This function creates two overview tables
\enumerate{
\item tbl_overview_culture_pest:
Indicates for the active ingredient(s) searched for, in which crops and against which pathogens
products with the active ingredient searched for are authorised.
In addition, the column "number_alternative_products" indicates how many
authorised alternative products (without the active substances searched for)
are authorised for the crop-pathogen combination. The following columns show
the number of authorised products (based on the W number) with the active substance(s) searched for.
\item tbl_alternative_products:
This table contains a list of alternative products for the combination culture/pest organism
that can replace products with the active substance(s) you are looking for.
}
}
\examples{
name_active_ingredient <- c("Cypermethrin", "Lambda-Cyhalothrin","Etofenprox", "Deltamethrin")


alternative_products(name_active_ingredient)
print(list_alternative_product)
}
