% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rates.R
\name{application_rate_g_per_ha}
\alias{application_rate_g_per_ha}
\title{Calculate application rates for active ingredients}
\usage{
application_rate_g_per_ha(
  product_uses,
  aggregation = c("max", "mean", "min"),
  dosage_units = c("percent_ww", "percent_vv", "state_of_matter"),
  skip_l_per_ha_without_g_per_L = TRUE,
  fix_l_per_ha = TRUE
)
}
\arguments{
\item{product_uses}{A tibble containing the columns 'wNbr', 'use_nr',
'application_area_de', 'min_dosage', 'max_dosage', 'min_rate', 'max_rate',
from the 'uses' table in a \link{srppp_dm} object, as well as the columns
'percent' and 'g_per_L' from the 'ingredients' table in a \link{srppp_dm} object.}

\item{aggregation}{How to represent a range if present, e.g. "max" (default)
or "mean".}

\item{dosage_units}{If no units are given, or units are "\%", then the applied
amount in g/ha is calculated using a reference application volume and the
dosage. As the dosage units are not explicitly given, we can specify our
assumptions about these using this argument (currently not implemented, i.e.
specifying the argument has no effect).}

\item{skip_l_per_ha_without_g_per_L}{Per default, uses where the use rate
has units of l/ha are skipped, if there is not product concentration
in g/L. This was also done in the 2023 indicator project.}

\item{fix_l_per_ha}{During the review of the 2023 indicator
project calculations, a number of cases were identified where the unit
l/ha specifies a water volume, and not a product volume. If TRUE (default),
these cases are corrected, if FALSE, these cases are discarded.}
}
\value{
A tibble containing one additional column 'rate_g_per_ha'
}
\description{
Calculate application rates for active ingredients
}
\note{
The reference application volume is used if there is no 'expenditure'.
It is selected only based on the product application area. This is not correct
if hops ('Hopfen') is the culture, as it has a unique reference application
volume of 3000 L/ha.
Applications to hops were excluded for calculating mean use rates in the
indicator project (Korkaric 2023), arguing that it is not grown in large
areas in Switzerland.
}
\examples{
\dontrun{
library(srppp)
library(dplyr, warn.conflicts = FALSE)
library(dm, warn.conflicts = FALSE)
srppp_cur <- srppp_dm()

product_uses_with_ingredients <- srppp_cur |>
  dm_filter(substances =
    (substance_de \%in\% c("Halauxifen-methyl", "Kupfer (als Kalkpr\u00E4parat)"))) |>
  dm_flatten_to_tbl(uses, products) |>
  left_join(srppp_cur$ingredients, by = join_by(pNbr),
    relationship = "many-to-many") |>
  left_join(srppp_cur$substances, by = join_by(pk)) |>
  select(wNbr, name, use_nr,
    min_dosage, max_dosage, min_rate, max_rate, units_de,
    application_area_de,
    substance_de, percent, g_per_L)

application_rate_g_per_ha(product_uses_with_ingredients) |>
  filter(name \%in\% c("Cerelex", "Pixxaro EC", "Bordeaux S")) |>
  select(ai = substance_de, app_area = application_area_de,
    min_d = min_dosage,  max_d = max_dosage,
    min_r = min_rate, max_r = max_rate,
    units_de, rate = rate_g_per_ha) |>
  print(n = Inf)
}
}
