% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{get_descriptions}
\alias{get_descriptions}
\title{Get descriptions from a node with childrent that hold descriptions}
\usage{
get_descriptions(node, code = FALSE)
}
\arguments{
\item{node}{The node to look at}

\item{code}{Do the description nodes have a child holding a code?}
}
\description{
Get descriptions from a node with childrent that hold descriptions
}
\keyword{internal}
