% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{get_great_grandparent_wNbr}
\alias{get_great_grandparent_wNbr}
\title{Get W-Number from grandparent node}
\usage{
get_great_grandparent_wNbr(node)
}
\description{
Get W-Number from grandparent node
}
\keyword{internal}
