% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{get_use_nr}
\alias{get_use_nr}
\title{Get use number from parent node (indication information node)}
\usage{
get_use_nr(node)
}
\description{
Get use number from parent node (indication information node)
}
\keyword{internal}
