% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product-rates.R
\name{product_rates}
\alias{product_rates}
\title{Calculate product application rates}
\usage{
product_rates(
  product_uses,
  aggregation = c("max", "mean", "min"),
  dosage_units = c("percent_ww", "percent_vv", "state_of_matter"),
  fix_l_per_ha = TRUE
)
}
\arguments{
\item{product_uses}{A tibble containing the columns 'pNbr', 'use_nr',
'application_area_de', 'min_dosage', 'max_dosage', 'min_rate', 'max_rate',
from the 'uses' table in a \link{srppp_dm} object, as well as the columns
'percent' and 'g_per_L' from the 'ingredients' table in a \link{srppp_dm} object and
'culture_de' from the 'cultures' table in \link{srppp_dm} object.}

\item{aggregation}{How to represent a range if present, e.g. "max" (default)
or "mean".}

\item{dosage_units}{If no units are given, or units are "\%", then the applied
amount in l/ha or kg/ha is calculated using a reference application volume and
the dosage. As the dosage units are not explicitly given, we can specify our
assumptions about these using this argument (currently not implemented, i.e.
specifying the argument has no effect).}

\item{fix_l_per_ha}{During the review of the 2023 indicator
project calculations, a number of cases were identified where the unit
l/ha specifies a water volume, and not a product volume. If TRUE (default),
these cases are corrected, if FALSE, these cases are discarded.}
}
\value{
A tibble containing one additional column 'rate_g_per_ha'
}
\description{
An application rate in l/ha or kg/ha is calculated from information
on dosage (product concentration in the application solution), application volume,
or directly from the product application rate. This is complicated by the fact
that a rate ("expenditure" in the XML file) with units l/ha can refer
to the application solution or to the liquid product.
}
\details{
In some cases (currently one), external information was found, indicating
that the "expenditure" is an application volume \link{l_per_ha_is_water_volume}.
}
\note{
A reference application volume is used if there is no 'expenditure'.
It is selected only based on the product application area and culture.
}
\examples{
\dontrun{
library(srppp)
library(dplyr, warn.conflicts = FALSE)
library(dm, warn.conflicts = FALSE)
sr <- srppp_dm()

product_uses <- sr$products |>
  filter(name == "BIOHOP AudiENZ") |>
  left_join(sr$uses, by = "pNbr",
            relationship = "many-to-many") |>
  left_join(sr$cultures, by = c("pNbr", "use_nr"),
            relationship = "many-to-many") |>
  left_join(sr$ingredients, by = c("pNbr"),
            relationship = "many-to-many") |>
  select(name, pNbr, use_nr,
    min_dosage, max_dosage, min_rate, max_rate, units_de,
    application_area_de, culture_de, pk, percent, g_per_L)

product_rates(product_uses, aggregation = "max") |>
  select(pNbr, name, culture_de, application_area_de,
  max_prod_rate=prod_rate, prod_unit) |>
  print(n = 10)
}
}
