% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-package.R
\docType{package}
\name{psmv-package}
\alias{psmv}
\alias{psmv-package}
\title{psmv: The Swiss Register of Plant Protection Products as an R Package}
\description{
Functions to obtain data objects from XML versions of the Swiss Register of Plant Protection Products 'PSMV'. An online version of the register can be accessed at \url{https://www.psm.admin.ch/de/produkte}.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://gitlab.agsad.admin.ch/psmwb/psmv/issues}
}

}
\author{
\strong{Maintainer}: Johannes Ranke \email{johannes.ranke@agroscope.admin.ch} (\href{https://orcid.org/0000-0003-4371-6538}{ORCID})

Other contributors:
\itemize{
  \item Marcel Mathis \email{marcel.mathis@agroscope.admin.ch} [contributor]
  \item Agroscope [copyright holder]
}

}
\keyword{internal}
