% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv_xml.R
\name{psmv_xml_clean_product_names}
\alias{psmv_xml_clean_product_names}
\title{Clean product names}
\usage{
psmv_xml_clean_product_names(names)
}
\arguments{
\item{names}{The product names that should be cleaned from comments}
}
\description{
This function is used in the data generation script for \link{psmv_xml_product_names}.
}
