% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv_xml.R
\name{psmv_xml_define_use_numbers}
\alias{psmv_xml_define_use_numbers}
\title{Define use identification numbers in a PSMV read in from an XML file}
\usage{
psmv_xml_define_use_numbers(psmv_xml = psmv_xml_get())
}
\arguments{
\item{psmv_xml}{An object as returned by 'psmv_xml_get'}
}
\value{
An psmv_xml object with use_nr added as an attribute of 'Indication' nodes.
}
\description{
Define use identification numbers in a PSMV read in from an XML file
}
\examples{
psmv_xml_define_use_numbers()
}
