% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv_xml.R
\name{psmv_xml_get}
\alias{psmv_xml_get}
\title{Read an XML version of the PSMV}
\usage{
psmv_xml_get(date = dplyr::last(names(fgpsm::psmv_xml_zip_files)))
}
\arguments{
\item{date}{The date of the publication in the format YYYY-MM-DD}
}
\value{
An object inheriting from 'psmv_xml', 'xml_document', 'xml_node'
}
\description{
Read an XML version of the PSMV
}
\examples{
psmv_xml <- psmv_xml_get()
print(psmv_xml)
class(psmv_xml)
psmv_xml_get("2023-04-04")
}
