% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_xml_get_ingredients}
\alias{psmv_xml_get_ingredients}
\title{Get ingredients for all products described in an XML version of the PSMV}
\usage{
psmv_xml_get_ingredients(psmv_xml = psmv_xml_get())
}
\arguments{
\item{psmv_xml}{An object as returned by 'psmv_xml_get'}
}
\description{
Get ingredients for all products described in an XML version of the PSMV
}
\examples{
psmv_xml_get_ingredients()
}
