% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_xml_get_parallel_imports}
\alias{psmv_xml_get_parallel_imports}
\title{Get Parallel Imports from an XML version of the PSMV}
\usage{
psmv_xml_get_parallel_imports(psmv_xml = psmv_xml_get())
}
\arguments{
\item{psmv_xml}{An object as returned by 'psmv_xml_get'}
}
\value{
A \link{tibble} with a row for each parallel import section
in the XML file.
}
\description{
Get Parallel Imports from an XML version of the PSMV
}
\examples{
library(psmv)
parallel_imports_2015 <- psmv_xml_get(2015) |>
  psmv_xml_get_parallel_imports()

# Get current list of parallel_imports
psmv_xml_get_parallel_imports()
}
