% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_xml_get_parallelimports}
\alias{psmv_xml_get_parallelimports}
\title{Get Parallel Imports from an XML version of the PSMV}
\usage{
psmv_xml_get_parallelimports(psmv_xml = psmv_xml_get())
}
\arguments{
\item{psmv_xml}{An object as returned by 'psmv_xml_get'}
}
\value{
A \link{tibble} with a row for each parallelimport section
in the XML file.
}
\description{
Get Parallel Imports from an XML version of the PSMV
}
\examples{
library(psmv)
parallelimports_2015 <- psmv_xml_get(2015) |>
  psmv_xml_get_parallelimports()

# Get current list of parallelimports
psmv_xml_get_parallelimports()
}
