% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv_xml.R
\name{psmv_xml_get_substances}
\alias{psmv_xml_get_substances}
\title{Get substances from an XML version of the PSMV}
\usage{
psmv_xml_get_substances(psmv_xml = psmv_xml_get())
}
\arguments{
\item{psmv_xml}{An object as returned by 'psmv_xml_get'}
}
\description{
Get substances from an XML version of the PSMV
}
\examples{
psmv_xml_get_substances()
}
