% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv_xml.R
\name{psmv_xml_get_uses}
\alias{psmv_xml_get_uses}
\title{Get uses ('indications') for all products described in an XML version of the PSMV}
\usage{
psmv_xml_get_uses(psmv_xml = psmv_xml_get())
}
\arguments{
\item{psmv_xml}{An object as returned by \link{psmv_xml_get} with use numbers
defined by \link{psmv_xml_define_use_numbers}}
}
\description{
Get uses ('indications') for all products described in an XML version of the PSMV
}
\examples{
psmv_xml <- psmv_xml_get()
psmv_xml <- psmv_xml_define_use_numbers(psmv_xml)
psmv_xml_get_uses(psmv_xml)
}
