% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-data.R
\docType{data}
\name{psmv_xml_idir}
\alias{psmv_xml_idir}
\title{Path to the directory where the published XML versions of the PSMV are stored}
\format{
length one character string
}
\usage{
psmv_xml_idir
}
\description{
This is a string, specifying the path to the input files, meant to be used on the RStudio server machines, where the
corresponding mount should be present.
}
\examples{
print(psmv_xml_idir)
}
\keyword{datasets}
