% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_xml_remove_duplicated_wNbrs}
\alias{psmv_xml_remove_duplicated_wNbrs}
\title{Remove product sections with duplicated W-Numbers}
\usage{
psmv_xml_remove_duplicated_wNbrs(
  psmv_xml = psmv_xml_get(),
  keep = c("last", "first")
)
}
\arguments{
\item{psmv_xml}{An object as returned by \link{psmv_xml_get}}

\item{keep}{How should we deal with product sections with identical W-Numbers?
The default strategy 'keep' only keeps the last of such sets of product sections.}
}
\value{
An 'psmv_xml' object with only product sections that have unique W-Numbers
}
\description{
Remove product sections with duplicated W-Numbers
}
\examples{
psmv_2015 <- psmv_xml_remove_duplicated_wNbrs(psmv_xml_get(2015))
}
