% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-use-rate.R
\docType{data}
\name{psmv_xml_units_convertible_to_g_per_ha}
\alias{psmv_xml_units_convertible_to_g_per_ha}
\title{Product application rate units convertible to grams active substance per hectare}
\format{
An object of class \code{character} of length 7.
}
\usage{
psmv_xml_units_convertible_to_g_per_ha
}
\description{
Product application rate units convertible to grams active substance per hectare
}
\examples{
library(psmv)
library(dplyr)
# These are the convertible units
psmv_xml_units_convertible_to_g_per_ha
}
\seealso{
\link{psmv_xml_application_rate_g_per_ha}
}
\keyword{datasets}
