% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml-data.R
\docType{data}
\name{psmv_xml_url}
\alias{psmv_xml_url}
\title{URL of the current XML version of the PSMV published by the BLV}
\format{
length one character string
}
\usage{
psmv_xml_url
}
\description{
URL of the current XML version of the PSMV published by the BLV
}
\examples{
print(psmv_xml_url)
}
\keyword{datasets}
