% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-package.R
\docType{package}
\name{srppp-package}
\alias{srppp}
\alias{srppp-package}
\title{srppp: Read the Swiss Register of Plant Protection Products}
\description{
Generate data objects from XML versions of the Swiss Register of Plant Protection Products. An online version of the register can be accessed at \url{https://www.psm.admin.ch/de/produkte}. There is no guarantee of correspondence of the data read in using this package with that online version, or with the original registration documents. Also, the Federal Food Safety and Veterinary Office, coordinating the authorisation of plant protection products in Switzerland, does not answer requests regarding this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agroscope-ch.github.io/srppp/}
  \item Report bugs at \url{https://github.com/agroscope-ch/srppp/issues}
}

}
\author{
\strong{Maintainer}: Johannes Ranke \email{johannes.ranke@agroscope.admin.ch} (\href{https://orcid.org/0000-0003-4371-6538}{ORCID})

Other contributors:
\itemize{
  \item Daniel Baumgartner (Provided feedback, explanations and background information) [contributor]
  \item Marcel Mathis \email{marcel.mathis@agroscope.admin.ch} [reviewer, contributor]
  \item Romualdus Kasteel \email{romualdus.kasteel@agroscope.admin.ch} (Provided feedback to version 0.3.4) [reviewer]
  \item Elisabeth Lutz \email{elisabeth.lutz@agroscope.admin.ch} [contributor]
  \item Agroscope [copyright holder]
}

}
\keyword{internal}
