% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{srppp_xml_clean_product_names}
\alias{srppp_xml_clean_product_names}
\title{Clean product names}
\usage{
srppp_xml_clean_product_names(names)
}
\arguments{
\item{names}{Character vector of product names that should be cleaned from comments}
}
\value{
Character vector of cleaned names
}
\description{
Clean product names
}
