% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{srppp_xml_get_ingredients}
\alias{srppp_xml_get_ingredients}
\title{Get ingredients for all products described in an XML version of the Swiss Register of Plant Protection Products}
\usage{
srppp_xml_get_ingredients(srppp_xml = srppp_xml_get())
}
\arguments{
\item{srppp_xml}{An object as returned by 'srppp_xml_get'}
}
\description{
Get ingredients for all products described in an XML version of the Swiss Register of Plant Protection Products
}
\examples{
\dontrun{
srppp_xml_get_ingredients()
}
}
