% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{srppp_xml_get_parallel_imports}
\alias{srppp_xml_get_parallel_imports}
\title{Get Parallel Imports from an XML version of the Swiss Register of Plant Protection Products}
\usage{
srppp_xml_get_parallel_imports(srppp_xml = srppp_xml_get())
}
\arguments{
\item{srppp_xml}{An object as returned by 'srppp_xml_get'}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with a row for each parallel import section
in the XML file.
}
\description{
Get Parallel Imports from an XML version of the Swiss Register of Plant Protection Products
}
\examples{
# Get current list of parallel_imports
\donttest{
srppp_xml_get_parallel_imports()
}
}
