% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{srppp_xml_get_uses}
\alias{srppp_xml_get_uses}
\title{Get uses ('indications') for all products described in an XML version of the Swiss Register of Plant Protection Products}
\usage{
srppp_xml_get_uses(srppp_xml = srppp_xml_get())
}
\arguments{
\item{srppp_xml}{An object as returned by \link{srppp_xml_get} with use numbers
defined by \link{srppp_xml_define_use_numbers}}
}
\value{
A \link[tibble:tibble]{tibble::tibble} of use definitions
}
\description{
Get uses ('indications') for all products described in an XML version of the Swiss Register of Plant Protection Products
}
\examples{
\donttest{
srppp_xml <- try(srppp_xml_get())
if (!inherits(srppp_xml, "try-error")) {
  srppp_xml <- srppp_xml_define_use_numbers(srppp_xml)
  srppp_xml_get_uses(srppp_xml)
}
}
}
