% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml-url.R
\name{srppp_xml_url}
\alias{srppp_xml_url}
\title{Determine the URL of the XML version of the Swiss Register of Plant Protection Products}
\usage{
srppp_xml_url()
}
\value{
length one character string
}
\description{
Determine the URL of the XML version of the Swiss Register of Plant Protection Products
}
\examples{
srppp_xml_url()
}
