% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rates.R
\docType{data}
\name{units_convertible_to_g_per_ha}
\alias{units_convertible_to_g_per_ha}
\title{Product application rate units convertible to grams active substance per hectare}
\format{
An object of class \code{character} of length 7.
}
\usage{
units_convertible_to_g_per_ha
}
\description{
Product application rate units convertible to grams active substance per hectare
}
\examples{
library(srppp)
library(dplyr)
# These are the convertible units
units_convertible_to_g_per_ha
}
\seealso{
\link{application_rate_g_per_ha}
}
\keyword{datasets}
