% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{description_table}
\alias{description_table}
\title{Get a table of descriptions for a certain Meta Information Tag}
\usage{
description_table(
  srppp_xml,
  tag_name,
  code = FALSE,
  latin = FALSE,
  parent_keys = FALSE
)
}
\description{
Get a table of descriptions for a certain Meta Information Tag
}
\keyword{internal}
