% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{get_descriptions}
\alias{get_descriptions}
\title{Get descriptions from a node with children that hold descriptions}
\usage{
get_descriptions(node, code = FALSE, latin = FALSE, parent_keys = FALSE)
}
\arguments{
\item{node}{The node to look at}

\item{code}{Do the description nodes have a child holding a code?}

\item{latin}{Are there latin descriptions (e.g. for pest descriptions)}

\item{parent_keys}{For culture descriptions, we also return up to two primary keys
that link to parent cultures}
}
\description{
Get descriptions from a node with children that hold descriptions
}
\keyword{internal}
