% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-rates.R
\docType{data}
\name{l_per_ha_is_water_volume}
\alias{l_per_ha_is_water_volume}
\title{Use definitions where the rate in l/ha refers to the volume of the spraying solution}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
l_per_ha_is_water_volume
}
\description{
Use definitions where the rate in l/ha refers to the volume of the spraying solution
}
\examples{
library(srppp)
l_per_ha_is_water_volume
}
\seealso{
\link{application_rate_g_per_ha}
}
\keyword{datasets}
