% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_dm}
\alias{psmv_dm}
\alias{print.psmv_dm}
\title{Create a dm object from an XML version of the PSMV}
\usage{
psmv_dm(from = psmv_xml_url, remove_duplicates = TRUE)

\method{print}{psmv_dm}(x, ...)
}
\arguments{
\item{from}{A specification of the way to retrieve the XML}

\item{remove_duplicates}{Should duplicates based on wNbrs be removed?}

\item{x}{A \link{psmv_dm} object}

\item{\dots}{Not used}
}
\value{
A \link{dm} object with tables linked by foreign keys
pointing to primary keys, i.e. with referential integrity.
}
\description{
Create a dm object from an XML version of the PSMV
}
\examples{
\dontrun{
library(dm)

psmv_cur <- psmv_dm()
dm_examine_constraints(psmv_cur)
dm_draw(psmv_cur)

# Show some information for products named 'Boxer'
psmv_cur |>
  dm_filter(products = (name == "Boxer")) |>
  dm_nrow()
}
}
