% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv_xml.R
\name{psmv_dm}
\alias{psmv_dm}
\title{Create a dm object from an XML version of the PSMV}
\usage{
psmv_dm(date = last(names(psmv::psmv_xml_zip_files)))
}
\arguments{
\item{date}{A number giving a year starting from 2011 up to the current year
or the date of the publication as a length one character vector in the
format YYYY-MM-DD}
}
\value{
A \link{dm} object with tables linked by foreign keys
pointing to primary keys, i.e. with referential integrity.
}
\description{
Create a dm object from an XML version of the PSMV
}
\examples{
psmv_2017 <- psmv_dm(2017)
dm_examine_constraints(psmv_2017)
# Show some information for products named 'Boxer'
psmv_2017 |>
  dm_filter(products = (name == "Boxer"))
}
