% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_xml_clean_product_names}
\alias{psmv_xml_clean_product_names}
\title{Clean product names}
\usage{
psmv_xml_clean_product_names(names)
}
\arguments{
\item{names}{The product names that should be cleaned from comments}
}
\description{
Clean product names
}
