% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_xml_get}
\alias{psmv_xml_get}
\title{Read an XML version of the PSMV}
\usage{
psmv_xml_get(date = last(names(psmv::psmv_xml_zip_files)))
}
\arguments{
\item{date}{A number giving a year starting from 2011 up to the current year
or the date of the publication as a length one character vector in the
format YYYY-MM-DD}
}
\value{
An object inheriting from 'psmv_xml', 'xml_document', 'xml_node'
}
\description{
Read an XML version of the PSMV
}
\examples{
psmv_2015 <- psmv_xml_get(2015)
print(psmv_2015)
class(psmv_2015)

# The current PSMV
psmv_xml <- psmv_xml_get()
}
