% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_xml_get}
\alias{psmv_xml_get}
\alias{psmv_xml_get.NULL}
\alias{psmv_xml_get.character}
\alias{psmv_xml_get_from_path}
\title{Read an XML version of the PSMV}
\usage{
psmv_xml_get(from, ...)

\method{psmv_xml_get}{`NULL`}(from, ...)

\method{psmv_xml_get}{character}(from, ...)

psmv_xml_get_from_path(path, from)
}
\arguments{
\item{from}{A specification of the way to retrieve the XML}

\item{\dots}{Unused argument introduced to facilitate future extensions}

\item{path}{A path to a zipped PSMV XML file}
}
\value{
An object inheriting from 'psmv_xml', 'xml_document', 'xml_node'
}
\description{
Read an XML version of the PSMV
}
\examples{
# The current PSMV as available from the FOAG website
psmv_cur <- psmv_xml_get()
# The current PSMV as available from the FOAG website
psmv_cur <- psmv_xml_get(psmv_xml_url)
}
