% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv_xml.R
\name{psmv_xml_get_ingredients}
\alias{psmv_xml_get_ingredients}
\title{Get ingredients for all products described in an XML version of the PSMV}
\usage{
psmv_xml_get_ingredients(psmv_xml = psmv_xml_get(), cores = 1)
}
\arguments{
\item{psmv_xml}{An object as returned by 'psmv_xml_get'}

\item{cores}{The number of cores to use in mclapply}
}
\description{
This function takes a while for execution, as it extracts the
information from a few thousands of products.
}
