% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmv-xml.R
\name{psmv_xml_get_products}
\alias{psmv_xml_get_products}
\title{Get Products from an XML version of the PSMV}
\usage{
psmv_xml_get_products(
  psmv_xml = psmv_xml_get(),
  verbose = TRUE,
  remove_duplicates = TRUE,
  keep = c("last", "first")
)
}
\arguments{
\item{psmv_xml}{An object as returned by 'psmv_xml_get'}

\item{verbose}{Should we give some feedback?}

\item{remove_duplicates}{Should duplicates based on wNbrs be removed?}

\item{keep}{Passed to \link{psmv_xml_remove_duplicated_wNbrs}}
}
\value{
A \link{tibble} with a row for each product section
in the XML file. An attribute 'duplicated_wNbrs' is
also returned, containing duplicated W-Numbers, if applicable,
or NULL.
}
\description{
Get Products from an XML version of the PSMV
}
\examples{
# Get current list of products
psmv_xml_get_products()
}
