% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{srppp_dm}
\alias{srppp_dm}
\alias{print.srppp_dm}
\title{Create a dm object from an XML version of the PSMV}
\usage{
srppp_dm(from = srppp_xml_url, remove_duplicates = TRUE)

\method{print}{srppp_dm}(x, ...)
}
\arguments{
\item{from}{A specification of the way to retrieve the XML}

\item{remove_duplicates}{Should duplicates based on wNbrs be removed?}

\item{x}{A \link{srppp_dm} object}

\item{\dots}{Not used}
}
\value{
A \link{dm} object with tables linked by foreign keys
pointing to primary keys, i.e. with referential integrity.
}
\description{
Create a dm object from an XML version of the PSMV
}
\examples{
\dontrun{
library(dm)

srppp_cur <- srppp_dm()
dm_examine_constraints(srppp_cur)
dm_draw(srppp_cur)

# Show some information for products named 'Boxer'
srppp_cur |>
  dm_filter(products = (name == "Boxer")) |>
  dm_nrow()
}
}
