% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{srppp_dm}
\alias{srppp_dm}
\alias{print.srppp_dm}
\title{Create a dm object from an XML version of the Swiss Register of Plant Protection Products}
\usage{
srppp_dm(from = srppp_xml_url, remove_duplicates = TRUE)

\method{print}{srppp_dm}(x, ...)
}
\arguments{
\item{from}{A specification of the way to retrieve the XML}

\item{remove_duplicates}{Should duplicates based on wNbrs be removed?}

\item{x}{A \link{srppp_dm} object}

\item{\dots}{Not used}
}
\value{
A \link{dm} object with tables linked by foreign keys
pointing to primary keys, i.e. with referential integrity.
}
\description{
Create a dm object from an XML version of the Swiss Register of Plant Protection Products
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(dm, warn.conflicts = FALSE)

sr <- srppp_dm()
dm_examine_constraints(sr)
\dontrun{
dm_draw(sr)

}

# Show ingredients for products named 'Boxer'
sr$products |>
  filter(name == "Boxer") |>
  left_join(sr$ingredients) |>
  left_join(sr$substances) |>
  select(wNbr, name, pNbr, isSalePermission, substance_de, g_per_L)

# Show authorised uses of the original product
boxer_uses <- sr$products |>
  filter(name == "Boxer", isSalePermission == "false") |>
  left_join(sr$uses) |>
  select(wNbr, name, use_nr,
    min_dosage, max_dosage, min_rate, max_rate, units_de,
    waiting_period, time_units_de, application_area_de)
print(boxer_uses)

# Show crop for use number 1
boxer_uses |>
  filter(use_nr == 1) |>
  left_join(sr$cultures, join_by(wNbr, use_nr)) |>
  select(use_nr, culture_de)

# Show target pests for use number 1
boxer_uses |>
  filter(use_nr == 1) |>
  left_join(sr$pests, join_by(wNbr, use_nr)) |>
  select(use_nr, pest_de)

# Show obligations for use number 1
boxer_uses |>
  filter(use_nr == 1) |>
  left_join(sr$obligations, join_by(wNbr, use_nr)) |>
  select(use_nr, obligation_de) |>
  knitr::kable() |>
  print()

# Show application comments for use number 1
boxer_uses |>
  filter(use_nr == 1) |>
  left_join(sr$application_comments, join_by(wNbr, use_nr)) |>
  select(use_nr, application_comment_de)
}
