% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml.R
\name{srppp_xml_define_use_numbers}
\alias{srppp_xml_define_use_numbers}
\title{Define use identification numbers in an SRPPP read in from an XML file}
\usage{
srppp_xml_define_use_numbers(srppp_xml = srppp_xml_get())
}
\arguments{
\item{srppp_xml}{An object as returned by \link{srppp_xml_get}}
}
\value{
An object of the same class, with 'use_nr' added as an attribute of 'Indication' nodes.
}
\description{
Define use identification numbers in an SRPPP read in from an XML file
}
\examples{
\donttest{
try(srppp_xml_define_use_numbers())
}
}
