% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srppp-xml-url.R
\docType{data}
\name{srppp_xml_url}
\alias{srppp_xml_url}
\title{URL of the XML version of the Swiss Register of Plant Protection Products}
\format{
length one character string
}
\usage{
srppp_xml_url
}
\description{
URL of the XML version of the Swiss Register of Plant Protection Products
}
\examples{
print(srppp_xml_url)
}
\keyword{datasets}
